requirements.txt
anyio
asgiref
cachetools
certifi
charset-normalizer
Django
djangorestframework
docstring_parser
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
gunicorn
h11
httpcore
httpx
httpx-sse
idna
jsonpatch
jsonpointer
langchain
langchain-core
langchain-google-vertexai
langchain-text-splitters
langchain-google-firestore
langsmith
numpy
orjson
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
PyYAML
requests
requests-toolbelt
rsa
setuptools
shapely
six
sniffio
SQLAlchemy
sqlparse
tenacity
typing-inspection
typing_extensions
tzdata
urllib3
validators
vertexai
wheel
zstandard